backend/requirements.txt
livekit
livekit-agents>=0.12.11,<1.0.0
livekit-plugins-openai>=0.10.17,<1.0.0
livekit-plugins-silero 
python-dotenv
livekit-api
flask[async]
flask 
flask-cors
uvicorn
